Move ext and fileops
