+++ lib/data/repo/app_state.dart
@@ -1,6 +1,6 @@
 import 'dart:async';
 
-import 'package:genshin_mod_manager/data/extension/path_op_string.dart';
+import 'package:genshin_mod_manager/data/helper/path_op_string.dart';
 import 'package:genshin_mod_manager/data/mapper/latest_stream.dart';
 import 'package:genshin_mod_manager/domain/repo/app_state.dart';
 import 'package:genshin_mod_manager/domain/repo/latest_stream.dart';
