+++ lib/data/repo/filesystem_watcher.dart
@@ -1,8 +1,8 @@
 import 'dart:async';
 import 'dart:io';
 
-import 'package:genshin_mod_manager/data/extension/path_op_string.dart';
-import 'package:genshin_mod_manager/data/io/fsops.dart';
+import 'package:genshin_mod_manager/data/helper/fsops.dart';
+import 'package:genshin_mod_manager/data/helper/path_op_string.dart';
 import 'package:genshin_mod_manager/data/mapper/latest_stream.dart';
 import 'package:genshin_mod_manager/domain/entity/fs_event.dart';
 import 'package:genshin_mod_manager/domain/entity/mod.dart';
