+++ lib/ui/route/home_shell/home_shell.dart
@@ -8,7 +8,7 @@ import 'package:flutter/foundation.dart';
 import 'package:flutter/gestures.dart';
 import 'package:flutter/scheduler.dart';
 import 'package:genshin_mod_manager/data/constant.dart';
-import 'package:genshin_mod_manager/data/extension/path_op_string.dart';
+import 'package:genshin_mod_manager/data/helper/path_op_string.dart';
 import 'package:genshin_mod_manager/data/repo/filesystem_watcher.dart';
 import 'package:genshin_mod_manager/data/repo/preset.dart';
 import 'package:genshin_mod_manager/domain/entity/mod_category.dart';
